+++ gnuradio-runtime/python/gnuradio/gru/CMakeLists.txt
@@ -9,6 +9,5 @@ include(GrPython)
 
 GR_PYTHON_INSTALL(FILES
     __init__.py
-    msgq_runner.py
     DESTINATION ${GR_PYTHON_DIR}/gnuradio/gru
 )
