+++ gnuradio-runtime/python/gnuradio/gru/__init__.py
@@ -1,4 +1,3 @@
 # make this a package
 
 # Import gru stuff
-from .msgq_runner import *
