+++ gr-uhd/apps/uhd_rx_nogui
@@ -51,11 +51,12 @@ blocks.
 import math
 import sys
 from argparse import ArgumentParser
-from gnuradio import gr, gru, uhd, audio
+from gnuradio import gr, uhd, audio
 from gnuradio import filter
 from gnuradio import analog
 from gnuradio import eng_notation, eng_arg
 from gnuradio.eng_option import eng_option
+import msgq_runner
 
 
 # (device_rate, channel_rate, audio_rate, channel_pass, channel_stop, demod)
@@ -191,7 +192,7 @@ class app_top_block(gr.top_block):
         if self.options.show_async_msg:
             self.async_msgq = gr.msg_queue(0)
             self.async_src = uhd.amsg_source("", self.async_msgq)
-            self.async_rcv = gru.msgq_runner(self.async_msgq, self.async_callback)
+            self.async_rcv = msgq_runner.msgq_runner(self.async_msgq, self.async_callback)
 
     def async_callback(self, msg):
         """ Callback for processing async messages """
